+++ src/indicators/r030.rs
@@ -47,11 +47,21 @@ impl Calculate for R030 {
                             if award_supplier_ids.contains(id) {
                                 set_result!(item, OCID, ocid, R030, 1.0);
                                 set_result!(item, Tenderer, id, R030, 1.0);
+                                if item.map {
+                                    item.maps.ocid_tenderer_r030.entry(ocid.to_owned()).or_default().insert(id.clone());
+                                }
                             }
                         }
                     }
                 }
             }
         }
+    }
+
+    fn reduce(&self, item: &mut Indicators, other: &mut Indicators) {
+        // If each OCID appears on only one line of the file, no overwriting will occur.
+        item.maps
+            .ocid_tenderer_r030
+            .extend(std::mem::take(&mut other.maps.ocid_tenderer_r030));
     }
 }